pyproject.toml
[tool.poetry]
name = "eduki_data_engineering"
version = "0.1.0"
description = ""
authors = ["Emmanouil (Manos) Palaiologos <manoscoffee@gmail.com>"]
readme = "README.md"
packages = [{include = "eduki_data_engineering"}]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-bigquery = "*"
pandas = "^2.2.2"
polars = "^0.20.21"
psycopg2-binary = "^2.9.9"
numpy = "^1.26.4"
pyyaml = "^6.0.1"
pandas-gbq = "^0.22.0"
tqdm = "^4.66.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"